Added All Improvements from Master

+++ NotEnoughEncodes/SmallScripts.cs
@@ -32,12 +32,31 @@ namespace NotEnoughEncodes
             try
             {
                 //Delete Files, because of lazy dump****
-                Directory.Delete("Chunks", true);
-                File.Delete("splitted.log");
-                File.Delete("encoded.txt");
-                Directory.Delete("AudioExtracted", true);
-                Directory.Delete("AudioEncoded", true);
-                File.Delete("no_audio.mkv");
+                if (File.Exists("splitted.log"))
+                {
+                    File.Delete("splitted.log");
+                }
+                if (File.Exists("encoded.txt"))
+                {
+                    File.Delete("encoded.txt");
+                }
+                if (File.Exists("no_audio.mkv"))
+                {
+                    File.Delete("no_audio.mkv");
+                }
+                if (Directory.Exists("AudioExtracted"))
+                {
+                    Directory.Delete("AudioExtracted", true);
+                }
+                if (Directory.Exists("AudioEncoded"))
+                {
+                    Directory.Delete("AudioEncoded", true);
+                }
+                if (Directory.Exists("Chunks"))
+                {
+                    Directory.Delete("Chunks", true);
+                }
+
             }
             catch { }
         }
@@ -45,11 +64,29 @@ namespace NotEnoughEncodes
         {
             try
             {
-                Directory.Delete(path + "\\Chunks", true);
-                File.Delete(path + "\\splitted.log");
-                Directory.Delete(path + "\\AudioExtracted", true);
-                Directory.Delete(path + "\\AudioEncoded", true);
-                File.Delete(path + "\\no_audio.mkv");
+
+                if (File.Exists(path + "\\splitted.log"))
+                {
+                    File.Delete(path + "\\splitted.log");
+                }
+
+                if (File.Exists(path + "\\no_audio.mkv"))
+                {
+                    File.Delete(path + "\\no_audio.mkv");
+                }
+                if (Directory.Exists(path + "\\AudioExtracted"))
+                {
+                    Directory.Delete(path + "\\AudioExtracted", true);
+                }
+                if (Directory.Exists(path + "\\AudioEncoded"))
+                {
+                    Directory.Delete(path + "\\AudioEncoded", true);
+                }
+                if (Directory.Exists(path + "\\Chunks"))
+                {
+                    Directory.Delete(path + "\\Chunks", true);
+                }
+
             }
             catch { }
 
@@ -81,9 +118,35 @@ namespace NotEnoughEncodes
         //Checks the dependencies (ffmpeg, aomenc, ffprobe)
         public static void CheckDependencies()
         {
-            bool aomencExist = File.Exists("aomenc.exe");
-            bool ffmpegExist = File.Exists("ffmpeg.exe");
-            bool ffprobeExist = File.Exists("ffprobe.exe");
+            bool aomencExist = false;
+            bool ffmpegExist = false;
+            bool ffprobeExist = false;
+            if (MainWindow.customAomencPathActive == true)
+            {
+                aomencExist = File.Exists(MainWindow.customAomencPath + "\\aomenc.exe");
+            }
+            else if (MainWindow.customAomencPathActive == false)
+            {
+                aomencExist = File.Exists("aomenc.exe");
+            }
+
+            if (MainWindow.customFfmpegPathActive == true)
+            {
+                ffmpegExist = File.Exists(MainWindow.customFfmpegPath + "\\ffmpeg.exe");
+            }
+            else if (MainWindow.customFfmpegPathActive == false)
+            {
+                ffmpegExist = File.Exists("ffmpeg.exe");
+            }
+            if (MainWindow.customFfprobePathActive == true)
+            {
+                ffprobeExist = File.Exists(MainWindow.customFfprobePath + "\\ffprobe.exe");
+            }
+            else if (MainWindow.customFfprobePathActive == false)
+            {
+                ffprobeExist = File.Exists("ffprobe.exe");
+            }
+
             if (aomencExist == false || ffmpegExist == false || ffprobeExist == false)
             {
                 MessageBox.Show("Couldn't find all depedencies: \n aomenc found: " + aomencExist + "\n ffmpeg found: " + ffmpegExist + " \n ffprobe found: " + ffprobeExist);
@@ -105,6 +168,7 @@ namespace NotEnoughEncodes
                 CreateNoWindow = true,
                 WindowStyle = ProcessWindowStyle.Hidden,
                 FileName = "cmd.exe",
+                WorkingDirectory = MainWindow.ffprobePath + "\\",
                 Arguments = "/C ffprobe.exe -i " + input + " -show_entries format=duration -v quiet -of csv=" + '\u0022' + "p=0" + '\u0022',
                 RedirectStandardError = true,
                 RedirectStandardOutput = true
